fix permission problems with overwriting files.
when copying an „other file“ and the file already exists, owned by another user
of the same group, stawebg crashes, because shutil.copy tries to chmod the file,
which is only allowed to the owner of the file, even if the other user has write
access through group membership. shutil.copyfile does not chmod and therefore
should fix this problem.